+++ src/main/resources/scheme-pg.sql
@@ -74,7 +74,7 @@ authority VARCHAR(50) NOT NULL
 
 -- 登录日志
 CREATE TABLE IF NOT EXISTS log_login (
-id int8 NOT NULL PRIMARY KEY,
+id bigint NOT NULL PRIMARY KEY,
 create_time timestamp(6),
 ip varchar(255),
 username varchar(255)
@@ -82,7 +82,7 @@ username varchar(255)
 
 -- 重置密码
 CREATE TABLE IF NOT EXISTS reset_pwd_info (
-id int8 NOT NULL primary key,
+id bigint NOT NULL primary key,
 expires_in timestamp(6),
 secret_key varchar(255),
 user_id int8,
@@ -91,7 +91,7 @@ valid bool NOT NULL
 
 -- 验证码
 CREATE TABLE IF NOT EXISTS veri_code (
-id int8 NOT null primary key,
+id bigint NOT null primary key,
 code int4,
 create_time int8,
 expires int4,
@@ -100,7 +100,7 @@ phone varchar(255)
 
 -- 二维码（用于扫码登录）
 CREATE TABLE IF NOT EXISTS qr_code (
-id int8 not null primary key,
+id bigint not null primary key,
 is_ok bool,
 sid varchar(255),
 token varchar(255),
@@ -109,9 +109,9 @@ username varchar(255)
 
 -- 新闻
 CREATE TABLE IF NOT EXISTS news_info (
-id int8 not null primary key,
+id bigint not null primary key,
 title varchar(255),
-user_id int8,
+user_id bigint,
 status int4,
 content text,
 image_url varchar(255),
@@ -121,9 +121,9 @@ tags varchar(255)
 
 -- 新闻标签
 CREATE TABLE IF NOT EXISTS news_tag (
-id int8 not null primary key,
-news_info_id int8,
-tag_id int8
+id bigint not null primary key,
+news_info_id bigint,
+tag_id bigint
 );
 
 -- 自增序列
